backend/requirements.txt
langchain
python-dotenv
streamlit
flask
streamlit
google-generativeai